window - push down static method to base
